requirements.txt
yolov9pip==0.0.4
huggingface_hub